removed dtype options

+++ image_generation.py
@@ -7,7 +7,7 @@ pipe = None
 
 def sd_model(model_id, dtype):
     global pipe
-    pipe = StableDiffusionPipeline.from_pretrained(model_id, torch_dtype=dtype, safety_checker = None, requires_safety_checker = False)
+    pipe = StableDiffusionPipeline.from_pretrained(model_id, safety_checker = None, requires_safety_checker = False)
     pipe = pipe.to(device)
 
 def generate_images(names, prompts, sd_folder_path1):
